requirements.txt
opencv-python~=4.5.2.54
Pillow~=8.2.0
pip~=21.1.3
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
Jinja2~=3.0.1
click~=8.0.1
itsdangerous~=2.0.1
numpy~=1.20.3
setuptools~=57.0.0
greenlet~=1.1.0
SQLAlchemy~=1.4.20
pytesseract~=0.3.7
Flask~=2.0.1